# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a basic ticket service written in go.",
        "title": "Go-Ticket API",
        "contact": {
            "name": "Max Greß",
            "url": "https://github.com/mgr1054/go-ticket",
            "email": "max.gress@student.reutlingen-university.de"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/secured/events": {
            "get": {
                "description": "Sends Array Of Events\nallowed: user, admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get All Events",
                "operationId": "get-events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Could not get events\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new Event\nallowed: admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create Event",
                "operationId": "create-event",
                "parameters": [
                    {
                        "description": "Create Event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.NewEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Could not create Event\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Could not create Event\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secured/events/{date}": {
            "get": {
                "description": "Sends Events for a Date\nallowed: user, admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get Event By Date",
                "operationId": "get-event-by-date",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Event not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secured/events/{id}": {
            "get": {
                "description": "Sends a Event with ID\nallowed: user, admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get Event By ID",
                "operationId": "get-event-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Event not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates Event with given ID\nallowed: admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Update Event By ID",
                "operationId": "update-event-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Event could not be updated with provided data\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Event not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Could not update Event\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes Event with given ID\nallowed: admin",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Delete Event By ID",
                "operationId": "delete-event-by-id",
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Event deleted\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Event not found!\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Could not create Event\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secured/events/{location}": {
            "get": {
                "description": "Sends Events for a Location\nallowed: user, admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get Event By Location",
                "operationId": "get-event-by-location",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Event not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.NewEvent": {
            "type": "object",
            "required": [
                "band_name",
                "capacity",
                "date",
                "location",
                "price"
            ],
            "properties": {
                "band_name": {
                    "type": "string",
                    "example": "Deichkind"
                },
                "capacity": {
                    "type": "integer",
                    "example": 35000
                },
                "date": {
                    "type": "string",
                    "example": "2022-10-11"
                },
                "location": {
                    "type": "string",
                    "example": "Olympiastadion"
                },
                "price": {
                    "type": "string",
                    "example": "55"
                }
            }
        },
...
        "models.Event": {
            "type": "object",
            "properties": {
                "band_name": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                }
            }
        },
...
    }
}